Add BC layer for generate method signature
